Replace Docker build-and-run script with .Net Core script
The Docker build-and-run.sh script was deleted and replaced with a new StartGolfClapBot.bat script that updates a Git repository and runs a .Net Core program. Additionally, a random delay was introduced before the bot sends messages in TwitchWorker.cs to reduce bot spamming. The delay varies between 8 and 12 seconds.